requirements.txt
flask>=3.0.0
mysql-connector-python>=9.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
anthropic>=0.40.0
